move queries to mutations for auth
