+++ packages/vulcan-users/lib/server/main.js
@@ -3,7 +3,8 @@ import './urls.js';
 import './graphql_context.js';
 import './callbacks.js';
 import './queries.js';
+import './mutations.js';
 
-export {default as createUser} from './create_user.js';
+export { default as createUser } from './create_user.js';
 export * from '../modules/index.js';
-export {default} from '../modules/index.js';
+export { default } from '../modules/index.js';
